[fix] compile error, moved folders to lib
