+++ lib/constants/operations.js
