+++ lib/core/harpeeConnect.js
@@ -1,4 +1,4 @@
-import util from "../../helper/utils";
+import util from "../helper/utils";
 
 import harpeeConnectConfig from "./harpeeConnectConfig";
