+++ lib/core/harpeeHttp.js
@@ -1,5 +1,5 @@
 import fetch from "node-fetch";
-import  util from "../../helpers/utils";
+import  util from "../helpers/utils";
 
 class HarpeeHttp {
    constructor(config = {}) {
