+++ lib/core/harpeeHttpAlt.js
@@ -1,5 +1,5 @@
 import fetch from "node-fetch";
-import util from "../../helpers/utils";
+import util from "../helpers/utils";
 
 //  harpeeHttpAlt is an alternative to harpeeHttp, it was used in places where `this` keyword was conflicting.
 /**
