+++ lib/core/harpeeModel.js
@@ -1,10 +1,10 @@
-import util from "../../helpers/utils";
+import util from "../helpers/utils";
 import HarpeeHttp from "./harpeeHttp";
 import callbackOrPromise from "./harpeeHttpAlt";
 import harpeeConnectConfig from "./harpeeConnectConfig";
 import harpeeModelConfig from "./harpeeModelConfig";
-import operations from "../../constants/operations";
-import validator from "../../helpers/validators";
+import operations from "../constants/operations";
+import validator from "../helpers/validators";
 import SqlHandler from "./sqlHandler" ;
 const sqlHandler = new SqlHandler();
