+++ lib/core/harpeeSchema.js
@@ -1,4 +1,4 @@
-import util from "../../helpers/utils";
+import util from "../helpers/utils";
 
 /** 
  * @typedef {Object} HarpeeSchemaObject
