+++ lib/core/harpeeUtilities.js
@@ -1,8 +1,8 @@
 import harpeeModelConfig from"./harpeeModelConfig";
 import harpeeConnectConfig from "./harpeeConnectConfig";
 import HarpeeHttp from "./harpeeHttp";
-import operations from "../../constants/operations";
-import util from"../../helpers/utils";
+import operations from "../constants/operations";
+import util from"../helpers/utils";
 
 /**
  * @callback responseCallback
