+++ lib/helpers/validateKeys.js
