+++ lib/helpers/validateKeysLength.js
