+++ lib/helpers/validateRequired.js
