+++ lib/helpers/validateTypes.js
