+++ lib/index.js
@@ -1,9 +1,9 @@
 import { createConnection, connect } from './core/harpeeConnect';
-import Schema from "./schema",
-import Model from "./model",
-import Utilities from "./utilities",
-import Logger from "./logger",
-import Sqler from "./sqler",
+import Schema from "./schema";
+import Model from "./model";
+import Utilities from "./utilities";
+import Logger from "./logger";
+import Sqler from "./sqler";
 export {
    createConnection,
    connect,
