+++ package.json
@@ -17,7 +17,7 @@
         "sql",
         "noSql"
     ],
-    "files":["constants","dist","helpers"],
+    "files":["dist"],
     "author": "Lucky Victory <luckyvictory54@gmail.com> (https://lucky-victory.dev)",
     "license": "MIT",
     "typings": "./types/index.d.ts",
